Make sure we call __hneg with half to remove ambigios error (#1736)
[ROCm/composable_kernel commit: 67497a044d450fbc0bcb099cfb0aa270cfb0aa6b]

+++ include/ck/utility/math_v2.hpp
@@ -611,7 +611,7 @@ inline __device__ int8_t neg<int8_t>(int8_t x)
 template <>
 inline __device__ half_t neg<half_t>(half_t x)
 {
-    return __hneg(x);
+    return __hneg(static_cast<__half>(x));
 };
 
 template <typename T>
